% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_articles.R
\name{add_articles}
\alias{add_articles}
\title{Add article identifiers to range shift data}
\usage{
add_articles(data)
}
\arguments{
\item{data}{range shift dataframe from get_shifts() function}
}
\value{
range shift dataframe supplemented with article identification information for each shift (author, doi)
}
\description{
Add article identifiers to range shift data
}
\examples{
get_shifts(group = "Birds", continent = "Africa") |> add_articles()
}
