% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_author_reported.R
\name{add_author_reported}
\alias{add_author_reported}
\title{Add author-reported shift results to range shifts}
\usage{
add_author_reported(data)
}
\arguments{
\item{data}{input shifts dataset from get_shifts() function.}
}
\value{
Range shifts dataset with added columns containing author-reported shift values, and the figure, table, or dataset in the original publication from which the shift values were pulled.
}
\description{
The BioShifts package standardizes author-reported data in several forms to rates of change in the latitudinal or elevational direction over time. This function supplements the calculated shift rates with shift findings as reported by original study authors.
}
\examples{
get_shifts() |> add_author_reported()
}
