% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cv.R
\name{add_cv}
\alias{add_cv}
\title{Add Climate Velocity to shifts dataframe}
\usage{
add_cv(
  data,
  type = "SA",
  stat = c("mean"),
  res = c(LAT = "25km", ELE = "1km"),
  suffix = F
)
}
\arguments{
\item{data}{Shifts dataframe from get_shifts() function}

\item{type}{Choice of climate velocity values from study area (SA) or species-specific study area (SP) polygons}

\item{stat}{Statistic of climate velocity to add. c("q25", "median", "mean", "q75", "sd").}

\item{res}{Spatial resolution with which climate velocities were calculated c("1km","25km","50km","110km"). Note that higher resolutions will generally have higher velocities, since climate velocity is calculated as climate trend / spatial gradient.}

\item{suffix}{Binary choice to add the resolution on to climate velocity variable columns. Use this if you plan to add multiple climate velocity resolutions to the same dataset.}
}
\value{
dataframe of range shifts supplemented with selected columns of climate velocity standardized to positive values in the poleward or elevational directions.
}
\description{
Add Climate Velocity to shifts dataframe
}
\examples{
get_shifts() |> add_cv(stat = c("mean"),res = c("LAT" = "25km", "ELE" = "1km")) |> dplyr::glimpse()
}
