% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_methods.R
\name{add_methods}
\alias{add_methods}
\title{Add methods to range shift documentations}
\usage{
add_methods(data)
}
\arguments{
\item{data}{input data from get_shifts().}
}
\value{
original shifts dataset, supplemented with methodological parameters for each individual range shift detection. See readme for details on methods columns.
}
\description{
add_methods supplements the minimal shifts database with methodological information describing the methods by which individal shifts were documented. These include variables such as sampling years, definitions of range parameters, and more.
}
\examples{
get_shifts() |> add_methods() |> dplyr::glimpse()
}
