% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_poly_info.R
\name{add_poly_info}
\alias{add_poly_info}
\title{Add polygon attributes to range shift dataframe}
\usage{
add_poly_info(data, type = "SA")
}
\arguments{
\item{data}{Range shifts dataframe from \code{get_shifts()} function}

\item{type}{Specification for study area polygon info ("SA"), or species-specific study area polygon info ("SP")}
}
\value{
Range shift dataframe supplemented with columns on spatial extent of study or species-specific polygons.
}
\description{
This function supplements range shift observations ()
}
\examples{
get_shifts() |> add_poly_info(type = "SA") |> dplyr::glimpse()
}
