% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_taxo.R
\name{add_taxo}
\alias{add_taxo}
\title{Add taxonomy to range shifts}
\usage{
add_taxo(data)
}
\arguments{
\item{df}{input shifts dataset from get_shifts() function.}
}
\value{
Range shifts dataset with added columns containing taxonomic classification information for every species name, as written in the original publication.
}
\description{
add_taxo() merges taxonomic classification to range shifts uploaded with the get_shifts() function.
}
\examples{
get_shifts() |> add_taxo()
}
