% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_trends.R
\name{add_trends}
\alias{add_trends}
\title{Add climate variable trends to range shift dataframe}
\usage{
add_trends(
  data,
  type = "SA",
  stat = c("mean"),
  exp = c("temp"),
  res = c(LAT = "25km", ELE = "1km")
)
}
\arguments{
\item{data}{dataframe of BioShifts range shifts from get_shifts() function}

\item{type}{Type of area over which trends are calculated: Article study areas ("SA"), or study areas cropped to species' range polygons ("SP").}

\item{stat}{Statistic of climate trends to add c("min", "1Q", "median", "mean", "3Q", "max").}

\item{exp}{Exposure variable c("temp","precip")}

\item{res}{Spatial resolution of climate grid cells with which the climate trends were calculated c("1km","25km","50km","110km"). Note that terrestrial latitudinal study areas are calculated at 1, 25, 50, 110km, marine latitudinal studies are calculated at 25, 50, 110km, and elevation studies are calculated only at 1km.}
}
\value{
range shift dataframe supplemented with selected trends in temperature (°C/year) or precipitation (inches ???!? / year) within study areas or species-specific study areas throughout the duration of the original study.
}
\description{
Add climate variable trends to range shift dataframe
}
\examples{
get_shifts() |> add_trends() |> dplyr::glimpse()
}
