% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_polygons.R
\name{download_polygons}
\alias{download_polygons}
\title{Download spatial data from OSF}
\usage{
download_polygons(
  type = "SA",
  polygon_folder = "./BioShiftR_polygons",
  timeout = 500
)
}
\arguments{
\item{type}{choice of study area ("SA") polygons, or species range polygons clipped to individual study areas ("SP"). Species range polygons will be more resolute in large study areas, but will take longer to download and use more disc space.}

\item{polygon_folder}{local directory in which to download polygon objects. Defaults to "./BioShiftR_polygons"}

\item{timeout}{timeout option if download fails. Increasing timeout may result in more stable downloads.}
}
\value{
data folder for polygon storage
}
\description{
BioShiftR relies on data from multiple sources. Spatial polygon datasets of all study areas, or species ranges within study areas are available on \href{https://osf.io/tp4hv/files/osfstorage}{Open Science Framework}, but need to be downloaded locally in order to use provided helper functions. This function only needs to be run once.
}
\examples{
\dontrun{
download_polyons(type = "SA")
}
}
