% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_articles.R
\name{add_articles}
\alias{add_articles}
\title{Add article identifiers to range shift data}
\usage{
add_articles(data)
}
\arguments{
\item{data}{Range shifts dataframe from the \code{get_shifts()} function}
}
\value{
Range shifts dataframe supplemented with article identification information for each shift: Author, DOI, and identifiers for the article within other datasets (BioShifts V1 and CoRE database of range shifts).
}
\description{
Add article identifiers to range shift data
}
\examples{
get_shifts(group = "Birds", continent = "Africa") |> add_articles() |> dplyr::glimpse()
}
