% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shifts.R
\name{get_shifts}
\alias{get_shifts}
\title{Get BioShifts Range Shifts - Start Here.}
\usage{
get_shifts(
  group = "All",
  realm = "All",
  continent = "All",
  type = c("LAT", "ELE")
)
}
\arguments{
\item{group}{Rough taxonomic subgroups for which to pull bioshifts data. Options are Algae, Birds, Fish, Fungi, Mammals, Marine Invertebrates, Nonvascular Plants, Reptiles and Amphibians, Terrestrial Invertebrates, and Vascular Plants, or All (default). This shortcut is meant to provide a coarse subsetting for data exploration, but for more precise taxonomic filtering, see add_taxo().}

\item{realm}{Subset of study realms for which to uplaod range shift data. Options are "Mar" (marine), "Ter" (terrestrial), or "All" (default).}

\item{continent}{Continent of studies for which to upload BioShifts data. Options include "North America", "South America", "Africa", "Europe", "Asia", "Oceania", "High Seas", or "All".}

\item{type}{Gradient over which to extract range shifts. Options are "ELE" for elevational shifts, or "LAT" for latitudinal shifts.}
}
\value{
Minimal data frame of calculated and author-reported species range shift values.
}
\description{
Get species' range shift values from the BioShifts database, filtered by taxon, study type or geography. BioShifts includes range shift observations of over 31,000 taxa within studies conducted around the world, published between
}
\examples{
get_shifts()
get_shifts(group = "Birds", continent = "Asia")
get_shifts(continent = c("North America","South America"), type = "ELE")
}
